+++ packages/tasks/lib/interfaces.ts
@@ -6,7 +6,9 @@ import {
 export type ITaskPriority = typeof TaskPriority[keyof typeof TaskPriority]
 export type ITaskStatus = typeof TaskStatus[keyof typeof TaskStatus]
 
-export interface ITaskExecute<TContext extends any | undefined> {
+export interface ITaskExecute<TContext = any, TPriority extends number = ITaskPriority> {
+  (context: TContext, tasks: ITaskEmitter<ITask<TContext, TPriority>>): Promise<any>
+  (context: TContext, tasks: ITaskEmitter<ITask<TContext, TPriority>>): any
   (context: TContext): Promise<any>
   (context: TContext): any
   (): Promise<any>
@@ -14,12 +16,12 @@ export interface ITaskExecute<TContext extends any | undefined> {
 }
 
 export interface ITask<
-  TTaskContext extends any | undefined = any,
+  TContext = any,
   TPriority extends number = ITaskPriority
 > {
   status: ITaskStatus
-  execute: ITaskExecute<TTaskContext>
-  context: TTaskContext
+  execute: ITaskExecute<TContext, TPriority>
+  context?: TContext
   id?: string
   priority?: TPriority
 }
@@ -30,9 +32,10 @@ export interface ITaskEmitterOptions<TTask> {
   onItemError?: (item: TTask, error: Error) => void
 }
 
-export interface ITaskEmitter<TTask extends ITask = ITask> {
+export interface ITaskEmitter<TTask extends ITask<any, number> = ITask> {
   readonly count: number
   readonly pendingCount: number
+  readonly runningCount: number
 
   add: (item: Omit<TTask, 'status'>) => void
   next: () => void